add command to update local repository instead of manually execute the command
modify the mechanism of update, add error report if failed and response on success

TIL bash EXPANSION

Failed to update. Probably due to the network connectivity issue, though it's just an assumption. The other possibilities are no running start-service program, uncommited changes on the git repository, Syntax error or Logic error on the javascript

having both stdout and stderr at the same time when executing a command is possible, even though it's unusual, but we still need to handle it regardless.
for example, the time command using stderr, so we need to redirect to stdout.